// src/contextualHelp/ui/common/redux/reducers/types.ts
export enum CommonActionType {
    ADD_AND_FOCUS_NEW_CELL = 'action.add_new_cell_and_focus_cell',
    INSERT_ABOVE_AND_FOCUS_NEW_CELL = 'action.insert_above_and_focus_cell',
    INSERT_BELOW_AND_FOCUS_NEW_CELL = 'action.insert_below_and_focus_cell',
    INSERT_ABOVE_FIRST_AND_FOCUS_NEW_CELL = 'action.insert_above_first_and_focus_cell',
    ADD_NEW_CELL = 'action.add_new_cell',
    ARROW_DOWN = 'action.arrow_down',
    ARROW_UP = 'action.arrow_up',
    CHANGE_CELL_TYPE = 'action.change_cell_type',
    CLICK_CELL = 'action.click_cell',
    CODE_CREATED = 'action.code_created',
    CONTINUE = 'action.continue',
    COPY_CELL_CODE = 'action.copy_cell_code',
    DELETE_CELL = 'action.delete_cell',
    EDITOR_LOADED = 'action.editor_loaded',
    EDIT_CELL = 'action.edit_cell',
    EXECUTE_ABOVE = 'action.execute_above',
    EXECUTE_ALL_CELLS = 'action.execute_all_cells',
    EXECUTE_CELL = 'action.execute_cell',
    EXECUTE_CELL_AND_ADVANCE = 'action.execute_cell_and_advance',
    EXECUTE_CELL_AND_BELOW = 'action.execute_cell_and_below',
    EXPORT = 'action.export',
    EXPORT_NOTEBOOK_AS = 'action.export_As',
    FOCUS_CELL = 'action.focus_cell',
    FOCUS_INPUT = 'action.focus_input',
    GET_VARIABLE_DATA = 'action.get_variable_data',
    GOTO_CELL = 'action.goto_cell',
    INSERT_ABOVE = 'action.insert_above',
    INSERT_ABOVE_FIRST = 'action.insert_above_first',
    INSERT_BELOW = 'action.insert_below',
    INTERRUPT_KERNEL = 'action.interrupt_kernel_action',
    LAUNCH_NOTEBOOK_TRUST_PROMPT = 'action.launch_notebook_trust_prompt',
    LOADED_ALL_CELLS = 'action.loaded_all_cells',
    LINK_CLICK = 'action.link_click',
    MOVE_CELL_DOWN = 'action.move_cell_down',
    MOVE_CELL_UP = 'action.move_cell_up',
    OPEN_SETTINGS = 'action.open_settings',
    PostOutgoingMessage = 'action.postOutgoingMessage',
    REDO = 'action.redo',
    REFRESH_VARIABLES = 'action.refresh_variables',
    RESTART_KERNEL = 'action.restart_kernel_action',
    RUN_BY_LINE = 'action.run_by_line',
    SAVE = 'action.save',
    SCROLL = 'action.scroll',
    SELECT_CELL = 'action.select_cell',
    SELECT_SERVER = 'action.select_server',
    SET_VARIABLE_EXPLORER_HEIGHT = 'action.set_variable_explorer_height',
    SET_VARIABLE_VIEW_HEIGHT = 'action.set_variable_view_height',
    SEND_COMMAND = 'action.send_command',
    SHOW_DATA_VIEWER = 'action.show_data_viewer',
    STEP = 'action.step',
    SUBMIT_INPUT = 'action.submit_input',
    TOGGLE_INPUT_BLOCK = 'action.toggle_input_block',
    TOGGLE_LINE_NUMBERS = 'action.toggle_line_numbers',
    TOGGLE_OUTPUT = 'action.toggle_output',
    TOGGLE_VARIABLE_EXPLORER = 'action.toggle_variable_explorer',
    UNDO = 'action.undo',
    UNFOCUS_CELL = 'action.unfocus_cell',
    UNMOUNT = 'action.unmount',
    VARIABLE_VIEW_LOADED = 'action.variable_view_loaded'
}